refactor: Change MessageDefinitionProvider to support multiple providers
Changed MessageDefinitionProvider to no longer have getMetaInfo
since it is a wide-open method with a not will defined interface.

This was replaced with the generateMessageIndex method since that
is the one use case that was being used. This also allows the
provider to be responsible for defining how its internal data
makes it to the MessageSink instead of forcing the caller
to have to know how to interpret the InputStream from getMetaInfo.