bug: Fix message recorder UDP connect issue
The server-side DatagramChannel was being bound only once
per environment.  This caused an issue when trying to connect
to a different destination port in sequential recordings. The
recording was only working for the first attempt and all further
attempts would fail to send the bytes to the destination.

The fix was to make a new channel and bind to the local port for
each recording attempt.