bug[ats_19ZD8]: Add ote server shutdown command
Eclipse 4.2 osgi close command does not wait for the
osgi framework to complete it's shutdown.  This means that
our server doesn't have a chance to properly clean up
so I add an 'x' command to shutdown the server which waits
for all bundles to stop running after shutting down the system
bundle.

Change-Id: Ia431eed1f16d3990674e1b4a9b39088d8f27065e