bug[ats_UM769]: New user creation causing the client to hang
When a new user is being persisted, the transaction calls into the
UserManager to get the current user for the author.  If this results in
a 2nd call into the currentUserProvider.get(), the thread will be blocked
indefinetely.  Thus, getCurrentUser() needs to use the cached, but not yet
persisted, user if called during user creation.

Change-Id: I45fdfb2ceb0c7437ccdf773a46fe00dcfd5ce6a4