bug[ats_08HWZ]: OSEE DSL needs attribute validation
The OSEE DSL grammar handles validating most things but cannot validate
the min/max occurances of attributes.  A new check was added to handle
this as well as ensuring a defaultValue is specified with min > 0.  Also
in this commit is a change to the grammar.  The change allows the optional
fields of an attribute type to be specified in any order.

Change-Id: I6261fa6753876584a30a53683f46553c81fb18ca