refinement[ats_ATS19824]: Add Branch Owner access level
Add new Branch Owner access level. This level allows
promoting/demoting of users between none and write,
in addition to having write access.

The branch access control of a child branch must be
copied from the parent branch at creation; the user
that created the branch must be assigned access
control of Branch Owner or above.

The branch access control dialog box must allow for
multiple select and access level change.  That is a
user with the appropriate access level must be able
to select multiple users and change the access level
of all selected to the same value.

Change-Id: Ic7660cbe015420cf412989f48a7a7da9e90373d0
Signed-off-by: mpotterc0k <marc.a.potter@boeing.com>